+++ RedPandaIDE/mainwindow.h
@@ -202,6 +202,7 @@ public slots:
     void onOJProblemCaseStarted(const QString& id, int current, int total);
     void onOJProblemCaseFinished(const QString& id, int current, int total);
     void onOJProblemCaseNewOutputGetted(const QString& id, const QString& line);
+    void onOJProblemCaseResetOutput(const QString& id, const QString& line);
     void cleanUpCPUDialog();
     void onDebugCommandInput(const QString& command);
     void onDebugEvaluateInput();